# Fix CI test performance - reduce execution time from 31min to <1min