@mojwang 2025-07-15
:tada: This PR is included in version 2.1.1 :tada:

The release is available on [GitHub release](https://github.com/mojwang/macbook-dev-setup/releases/tag/v2.1.1)

Your **[semantic-release](https://github.com/semantic-release/semantic-release)** bot :package::rocket: